requirements.txt
# Core
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0

# Deep Learning
tensorflow==2.14.0
keras==2.14.0

# Computer Vision
opencv-python==4.8.0.76
opencv-contrib-python==4.8.0.76
mtcnn==0.1.1

# Utils
scikit-learn==1.3.0
scipy==1.11.2
tqdm==4.66.1
python-dotenv==1.0.0

# Visualization
tensorboard==2.14.0
jupyter==1.0.0
ipykernel==6.25.1

# API
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0